refactor(bundle-size): extract getTabsterContext/getRoot from RootAPI class
Move the three static methods (getTabsterContext, getRoot, getRootByUId —
last one was dead code) out of the RootAPI class body and into a new
src/Context.ts. The 30+ internal call sites now import a free function
instead of going through `RootAPI.getTabsterContext(...)` on a class
that ships ~310 lines of unrelated instance code.

The RootAPI class still ships (it's the runtime backing for tabster.root),
but consumer modules that only want to walk the context tree no longer
pull the entire class definition through the import graph.

_autoRoot and _autoRootCreate are no longer `private` — Context.ts
reaches into them for the auto-root fallback. Marked /** @internal */.

Bundle-size:
- createTabster (core): 33.80 → 33.38 kB (-418 B)
- all exports: 96.47 → 95.40 kB (-1.07 kB)
- Each getX fixture: -350 to -560 B
