+++ src/Modalizer.ts
@@ -6,7 +6,7 @@
 import { nativeFocus } from "keyborg";
 import { _findFocusable } from "./Focusable.js";
 import { getTabsterOnElement } from "./Instance.js";
-import { RootAPI } from "./Root.js";
+import { getRoot, getTabsterContext } from "./Context.js";
 import { FocusedElementState } from "./State/FocusedElement.js";
 import { Keys } from "./Keys.js";
 import type * as Types from "./Types.js";
@@ -85,17 +85,14 @@ function createModalizerDummyManager(
 
     manager.setHandlers((dummyInput: DummyInput, isBackward: boolean) => {
         const el = element.get();
-        const container = el && RootAPI.getRoot(tabster, el)?.getElement();
+        const container = el && getRoot(tabster, el)?.getElement();
         const input = dummyInput.input;
         let toFocus: HTMLElement | null | undefined;
 
         if (container && input) {
             const dummyContainer = getDummyInputContainer(input);
 
-            const ctx = RootAPI.getTabsterContext(
-                tabster,
-                dummyContainer || input
-            );
+            const ctx = getTabsterContext(tabster, dummyContainer || input);
 
             if (ctx) {
                 toFocus = FocusedElementState.findNextTabbable(
@@ -254,8 +251,7 @@ export class Modalizer
         let uncontrolled: HTMLElement | null | undefined;
 
         const container =
-            currentElement &&
-            RootAPI.getRoot(tabster, currentElement)?.getElement();
+            currentElement && getRoot(tabster, currentElement)?.getElement();
 
         if (container) {
             const findProps: Types.FindNextProps = {
@@ -528,7 +524,7 @@ export function createModalizerAPI(
         const focusedElement = tabster.focusedElement.getFocusedElement();
         const focusedElementModalizer =
             focusedElement &&
-            RootAPI.getTabsterContext(tabster, focusedElement)?.modalizer;
+            getTabsterContext(tabster, focusedElement)?.modalizer;
 
         if (
             !focusedElement ||
@@ -539,7 +535,7 @@ export function createModalizerAPI(
             return;
         }
 
-        const ctx = RootAPI.getTabsterContext(tabster, outsideElement);
+        const ctx = getTabsterContext(tabster, outsideElement);
         const modalizer = ctx?.modalizer;
         const activeIdLocal = api.activeId;
 
@@ -594,8 +590,7 @@ export function createModalizerAPI(
         detail: Types.FocusedElementDetail
     ): void => {
         const ctx =
-            focusedElement &&
-            RootAPI.getTabsterContext(tabster, focusedElement);
+            focusedElement && getTabsterContext(tabster, focusedElement);
 
         // Modalizer behaviour is opt in, only apply to elements that have a tabster context
         if (!ctx || !focusedElement) {
@@ -638,8 +633,7 @@ export function createModalizerAPI(
                 const parentElement = tabster.getParent(focusedElement);
                 const parentModalizer =
                     parentElement &&
-                    RootAPI.getTabsterContext(tabster, parentElement)
-                        ?.modalizer;
+                    getTabsterContext(tabster, parentElement)?.modalizer;
 
                 if (parentModalizer) {
                     modalizer = parentModalizer;
@@ -746,7 +740,7 @@ export function createModalizerAPI(
         const element = tabster.focusedElement.getFocusedElement();
 
         if (element) {
-            const ctx = RootAPI.getTabsterContext(tabster, element);
+            const ctx = getTabsterContext(tabster, element);
             const modalizer = ctx?.modalizer;
 
             if (
@@ -955,10 +949,7 @@ export function createModalizerAPI(
             noFocusFirst?: boolean,
             noFocusDefault?: boolean
         ): boolean {
-            const ctx = RootAPI.getTabsterContext(
-                tabster,
-                elementFromModalizer
-            );
+            const ctx = getTabsterContext(tabster, elementFromModalizer);
 
             const modalizer = ctx?.modalizer;
 
@@ -1011,10 +1002,8 @@ export function createModalizerAPI(
         ): boolean {
             const modalizerToActivate: Types.Modalizer | undefined =
                 modalizerElementOrContainer
-                    ? RootAPI.getTabsterContext(
-                          tabster,
-                          modalizerElementOrContainer
-                      )?.modalizer
+                    ? getTabsterContext(tabster, modalizerElementOrContainer)
+                          ?.modalizer
                     : undefined;
 
             if (!modalizerElementOrContainer || modalizerToActivate) {
